requirements.txt
streamlit
chromadb
llama-index-core
llama-index-embeddings-huggingface
llama-index-llms-azure-inference
llama-index-vector-stores-chroma
llama-index-readers-file
azure-ai-inference
transformers
fastapi
uvicorn
gunicorn
llama-index
pypdf
sentence-transformers
azure-ai-ml
requests
python-dotenv
huggingface_hub
tiktoken